parallel-thinking-main/SFT/requirements.txt
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.24.0
peft>=0.7.0
trl>=0.7.0
datasets>=2.14.0
evaluate>=0.4.0
bitsandbytes>=0.41.0
scipy
scikit-learn
sentencepiece
protobuf
tiktoken
Pillow
matplotlib
tqdm
packaging
pyyaml
numpy
pandas
